Cleaned up project files after refactor.
