--- ArmEmu/TestSystem.cpp
@@ -1,120 +1,0 @@
-//! @file ArmEmu/TestSystem.cpp
-//! @brief The definition of an implementation of IArmSystem used for testing.
-//! @author GiantRobotLemur@na-se.co.uk
-//! @date 2022-2023
-//! @copyright This file is part of the Mighty Oak project which is released
-//! under LGPL 3 license. See LICENSE file at the repository root or go to
-//! https://github.com/GiantRobotLemur/MightyOak for full license details.
-////////////////////////////////////////////////////////////////////////////////
-
-////////////////////////////////////////////////////////////////////////////////
-// Header File Includes
-////////////////////////////////////////////////////////////////////////////////
-#include "Ag/Core/Exception.hpp"
-
-#include "TestSystem.hpp"
-
-namespace Ag {
-namespace Arm {
-
-////////////////////////////////////////////////////////////////////////////////
-// TestSystem Member Definitions
-////////////////////////////////////////////////////////////////////////////////
-//! @brief Constructs an emulated system for testing purposes.
-TestSystem::TestSystem() :
-    _resources(&_processor)
-{
-    // Cross link the processor with its system resources.
-    _processor.initialiseSystemResources(&_resources);
-}
-
-//! @brief Gets the resources associated with the emulated system.
-SystemResources &TestSystem::getResources()
-{
-    return _resources;
-}
-
-//! @brief Gets the object representing the emulated ARM processor.
-InstructionPipeline &TestSystem::getProcessor()
-{
-    return _processor;
-}
-
-// Inherited from IArmSystem.
-ProcessorMode TestSystem::getMode() const
-{
-    return _processor.getMode();
-}
-
-// Inherited from IArmSystem.
-uint32_t TestSystem::getCoreRegister(CoreRegister id) const
-{
-    uint32_t value = 0;
-
-    if (id <= CoreRegister::R15)
-    {
-        value = _processor.getRd(fromScalar<GeneralRegister>(toScalar(id)));
-    }
-    else if (id == CoreRegister::CPSR)
-    {
-        value = _processor.getPSR();
-    }
-    else if (id == CoreRegister::PC)
-    {
-        value = _processor.getPC();
-    }
-
-    return value;
-}
-
-// Inherited from IArmSystem.
-void TestSystem::setCoreRegister(CoreRegister id, uint32_t value)
-{
-    if (id <= CoreRegister::R15)
-    {
-        _processor.setRd(fromScalar<GeneralRegister>(toScalar(id)),
-                         value, true);
-    }
-    else if (id == CoreRegister::CPSR)
-    {
-        _processor.setPSR(value);
-    }
-    else if (id == CoreRegister::SPSR)
-    {
-        throw NotSupportedException("Setting SPSR on an ARMv2 processor");
-    }
-    else if (id == CoreRegister::PC)
-    {
-        _processor.setRn(GeneralRegister::R15, value);
-    }
-}
-
-// Inherited from IArmSystem.
-uint32_t TestSystem::readFromLogicalAddress(uint32_t logicalAddr, uint32_t length,
-                                            void *buffer) const
-{
-    return _processor.rawReadLogicalMemory(logicalAddr, buffer, length);
-}
-
-// Inherited from IArmSystem.
-void TestSystem::writeToLogicalAddress(uint32_t logicalAddr, uint32_t length,
-                                       const void *buffer)
-{
-    _processor.rawWriteLogicalMemory(logicalAddr, buffer, length);
-}
-
-// Inherited from IArmSystem.
-uint64_t TestSystem::run()
-{
-    return _processor.runPipeline(false);
-}
-
-// Inherited from IArmSystem.
-uint64_t TestSystem::runSingleStep()
-{
-    return _processor.runPipeline(true);
-}
-
-}} // namespace Ag::Arm
-////////////////////////////////////////////////////////////////////////////////
-
